refactor: remove traits and introduce extra crates
