+++ crates/pixi_install_pypi/src/lib.rs
@@ -4,7 +4,6 @@ use std::{
     sync::Arc,
 };
 
-use async_trait::async_trait;
 use chrono::{DateTime, Utc};
 use conda_pypi_clobber::PypiCondaClobberRegistry;
 use fancy_display::FancyDisplay;
@@ -16,6 +15,7 @@ use pixi_manifest::{
     pypi::pypi_options::{NoBinary, NoBuild, NoBuildIsolation},
 };
 use pixi_progress::await_in_progress;
+use pixi_python_status::PythonStatus;
 use pixi_record::PixiRecord;
 use pixi_utils::prefix::Prefix;
 use pixi_uv_context::UvResolutionContext;
@@ -30,6 +30,7 @@ use pypi_modifiers::{
 };
 use rattler_conda_types::Platform;
 use rattler_lock::{PypiIndexes, PypiPackageData, PypiPackageEnvironmentData};
+use rayon::prelude::*;
 use utils::elapsed;
 use uv_auth::store_credentials_from_url;
 use uv_client::{Connectivity, FlatIndexClient, RegistryClient, RegistryClientBuilder};
@@ -53,20 +54,111 @@ use pixi_reporters::{UvReporter, UvReporterOptions};
 
 pub type PyPIRecords = (PypiPackageData, PypiPackageEnvironmentData);
 
-#[async_trait]
-pub trait PythonEnvironmentProvider: Send + Sync {
-    async fn python_info_for_update(
-        &self,
-        prefix: &Prefix,
-        planned_records: &[PyPIRecords],
-    ) -> miette::Result<Option<rattler::install::PythonInfo>>;
-}
-
 pub(crate) mod conda_pypi_clobber;
 pub(crate) mod conversions;
 pub(crate) mod install_wheel;
 pub(crate) mod plan;
 pub(crate) mod utils;
+
+/// Continue or skip a PyPI prefix update based on the interpreter state.
+pub enum ContinuePyPIPrefixUpdate<'a> {
+    /// Continue with the update using the provided interpreter metadata.
+    Continue(&'a rattler::install::PythonInfo),
+    /// Skip the update entirely.
+    Skip,
+}
+
+/// Remove site-packages installed for an outdated interpreter so the next run
+/// starts from a clean slate.
+async fn uninstall_outdated_site_packages(site_packages: &Path) -> miette::Result<()> {
+    let mut dist_dirs = Vec::new();
+    for entry in fs_err::read_dir(site_packages).into_diagnostic()? {
+        let entry = entry.into_diagnostic()?;
+        if entry.file_type().into_diagnostic()?.is_dir() {
+            dist_dirs.push(entry.path());
+        }
+    }
+
+    let installed = dist_dirs
+        .into_par_iter()
+        .flat_map(|path| {
+            let installer_path = path.join("INSTALLER");
+            match fs_err::metadata(&installer_path) {
+                Ok(metadata) if metadata.len() as usize != consts::PIXI_UV_INSTALLER.len() => {
+                    return None;
+                }
+                Err(err) if err.kind() == std::io::ErrorKind::NotFound => {
+                    return None;
+                }
+                Ok(_) | Err(_) => {}
+            }
+
+            let Ok(Some(installed_dist)) = InstalledDist::try_from_path(&path) else {
+                return None;
+            };
+
+            let Ok(installer) = installed_dist.installer() else {
+                tracing::warn!(
+                    "could not get installer for {}: will not remove distribution",
+                    installed_dist.name()
+                );
+                return None;
+            };
+
+            if installer.unwrap_or_default() == consts::PIXI_UV_INSTALLER {
+                Some(installed_dist)
+            } else {
+                None
+            }
+        })
+        .collect::<Vec<_>>();
+
+    for dist_info in installed {
+        uv_installer::uninstall(&dist_info)
+            .await
+            .expect("uninstallation of old site-packages failed");
+    }
+
+    Ok(())
+}
+
+/// React on interpreter changes before running the PyPI updater. This may
+/// trigger cleanup of outdated site-packages or skip the update entirely.
+pub async fn on_python_interpreter_change<'a>(
+    status: &'a PythonStatus,
+    prefix: &Prefix,
+    pypi_records: &[PyPIRecords],
+) -> miette::Result<ContinuePyPIPrefixUpdate<'a>> {
+    match status {
+        PythonStatus::Removed { old } => {
+            let site_packages_path = prefix.root().join(&old.site_packages_path);
+            if site_packages_path.exists() {
+                uninstall_outdated_site_packages(&site_packages_path).await?;
+            }
+            Ok(ContinuePyPIPrefixUpdate::Skip)
+        }
+        PythonStatus::Changed { old, new } => {
+            if old.site_packages_path != new.site_packages_path {
+                let site_packages_path = prefix.root().join(&old.site_packages_path);
+                if site_packages_path.exists() {
+                    uninstall_outdated_site_packages(&site_packages_path).await?;
+                }
+            }
+            Ok(ContinuePyPIPrefixUpdate::Continue(new))
+        }
+        PythonStatus::Unchanged(info) | PythonStatus::Added { new: info } => {
+            if pypi_records.is_empty() {
+                let site_packages_path = prefix.root().join(&info.site_packages_path);
+                if site_packages_path.exists() {
+                    uninstall_outdated_site_packages(&site_packages_path).await?;
+                }
+                return Ok(ContinuePyPIPrefixUpdate::Skip);
+            }
+            Ok(ContinuePyPIPrefixUpdate::Continue(info))
+        }
+        PythonStatus::DoesNotExist => Ok(ContinuePyPIPrefixUpdate::Skip),
+    }
+}
 
 /// Configuration for PyPI environment updates, grouping basic environment settings
 pub struct PyPIUpdateConfig<'a> {
@@ -111,11 +203,10 @@ struct UvInstallerConfig {
 
 /// High-level interface for PyPI environment updates that handles all complexity internally
 /// This is full of lifetime, because internal uv datastructs require it.
-pub struct PyPIEnvironmentUpdater<'a, P: PythonEnvironmentProvider> {
+pub struct PyPIEnvironmentUpdater<'a> {
     config: PyPIUpdateConfig<'a>,
     build_config: PyPIBuildConfig<'a>,
     context_config: PyPIContextConfig<'a>,
-    python_env_provider: &'a P,
     // Names that should never be marked as extraneous in PyPI planning
     ignored_extraneous: HashSet<PackageName>,
 }
@@ -127,22 +218,17 @@ pub struct SeparatedDistributions {
     pub no_build_isolation_dists: Vec<(uv_distribution_types::Dist, InstallReason)>,
 }
 
-impl<'a, P> PyPIEnvironmentUpdater<'a, P>
-where
-    P: PythonEnvironmentProvider,
-{
+impl<'a> PyPIEnvironmentUpdater<'a> {
     /// Create a new PyPI environment updater with the given configurations
     pub fn new(
         config: PyPIUpdateConfig<'a>,
         build_config: PyPIBuildConfig<'a>,
         context_config: PyPIContextConfig<'a>,
-        python_env_provider: &'a P,
     ) -> Self {
         Self {
             config,
             build_config,
             context_config,
-            python_env_provider,
             ignored_extraneous: Default::default(),
         }
     }
@@ -161,18 +247,17 @@ where
     /// Update PyPI packages in the environment, handling all setup, planning, and execution
     pub async fn update(
         &self,
+        python_status: &PythonStatus,
         pixi_records: &[PixiRecord],
         pypi_records: &[PyPIRecords],
     ) -> miette::Result<()> {
-        // Determine global site-packages status
-        let python_info = match self
-            .python_env_provider
-            .python_info_for_update(self.config.prefix, pypi_records)
-            .await?
-        {
-            Some(info) => info,
-            None => return Ok(()),
-        };
+        let python_info =
+            match on_python_interpreter_change(python_status, self.config.prefix, pypi_records)
+                .await?
+            {
+                ContinuePyPIPrefixUpdate::Continue(info) => info,
+                ContinuePyPIPrefixUpdate::Skip => return Ok(()),
+            };
 
         // Install and/or remove python packages
         await_in_progress(
