+++ crates/pixi_python_status/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+name = "pixi_python_status"
+readme.workspace = true
+repository.workspace = true
+version = "0.1.0"
+
+[dependencies]
+rattler = { workspace = true }
+rattler_conda_types = { workspace = true }
